backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
robotexclusionrulesparser==1.7.1
apscheduler==3.10.4
python-multipart==0.0.20
